requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
twilio==9.10.1
python-dotenv==1.0.0
openai==1.12.0
